Iteration 1. Complete with role validation
