+++ starter-code/views/boss/newemployee.hbs
@@ -1,4 +1,4 @@
-<form action="/auth/newemployee" method="POST">
+<form action="/newemployee" method="POST">
   <input type="text" name="username" placeholder="Username">
   <input type="text" name="role" placeholder="Rol">
   <input type="submit" value="Agregar usuario">
