+++ starter-code/views/boss/profileboss.hbs
@@ -1,0 +1,8 @@
+<h1>Profile</h1>
+<h2>{{username}}</h2>
+
+<a href="/list">Listar minions |</a>
+<a href="/newemployee">Agregar minion |</a>
+<a href="/deleteemployee">Borrar minion |</a>
+<a href="/profile">Ir a Perfil |</a>
+<a href="/auth/logout">logout</a>
