--- starter-code/views/profileboss.hbs
@@ -1,5 +1,0 @@
-<h1>Profile</h1>
-<h2>{{username}}</h2>
-
-<a href="/auth/newemployee">Agregar minion</a>
-{{!-- <a href="/deleteemployee">Borrar minion</a> --}}
