+++ starter-code/views/ta/profile.hbs
@@ -1,0 +1,4 @@
+<h1>Profile</h1>
+<h2>{{username}}</h2>
+
+<a href="/auth/logout">logout</a>
